chore: converge repo structure to canonical pattern
Standardize configs, docs, and CI to match the canonical
Obsidian plugin pattern defined across all philoserf plugins.

- Replace AGENTS.md/symlink with standalone CLAUDE.md
- Remove repo-specific .gitignore entries
- Consolidate src/__mocks__/ into src/test-preload.ts with bunfig.toml
- Replace version-bump.ts with canonical version
- Add bun test to CI workflow
