+++ .gitignore
@@ -28,6 +28,3 @@ Thumbs.db
 
 # Planning
 .planning/
-
-# Obsidian
-.obsidian/
