--- src/__mocks__/preload.ts
@@ -1,3 +1,0 @@
-import { mock } from "bun:test";
-
-mock.module("obsidian", () => require("./obsidian"));
